a bit more refactoring
1. to align with the pattern used for Functions convenience
2. some more JavaDoc
